+++ src/main/java/com/kucw/security/rowmapper/ProductRowMapper.java
@@ -19,7 +19,7 @@ public class ProductRowMapper implements RowMapper {
         product.setProductName(rs.getString("product_name"));
         product.setCategory(ProductCategory.valueOf(rs.getString("category")));
         product.setImageUrl(rs.getString("image_url"));
-        product.setPrice(rs.getInt("price"));
+        product.setPrice(rs.getBigDecimal("price"));
         product.setStock(rs.getInt("stock"));
         product.setDescription(rs.getString("description"));
         product.setCreatedDate(rs.getTimestamp("created_date"));
